+++ src/mixins/java/org/spongepowered/common/mixin/api/mcp/world/server/ServerWorldMixin_API.java
@@ -26,6 +26,7 @@ package org.spongepowered.common.mixin.api.mcp.world.server;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.block.Block;
@@ -59,10 +60,13 @@ import org.spongepowered.api.event.SpongeEventFactory;
 import org.spongepowered.api.event.world.ExplosionEvent;
 import org.spongepowered.api.fluid.FluidType;
 import org.spongepowered.api.scheduler.ScheduledUpdateList;
+import org.spongepowered.api.util.TemporalUnits;
 import org.spongepowered.api.world.ChunkRegenerateFlag;
 import org.spongepowered.api.world.ServerLocation;
 import org.spongepowered.api.world.storage.WorldProperties;
 import org.spongepowered.api.world.storage.WorldStorage;
+import org.spongepowered.api.world.weather.Weather;
+import org.spongepowered.api.world.weather.Weathers;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -70,6 +74,7 @@ import org.spongepowered.asm.util.PrettyPrinter;
 import org.spongepowered.common.SpongeCommon;
 import org.spongepowered.common.accessor.world.raid.RaidManagerAccessor;
 import org.spongepowered.common.accessor.world.storage.SaveHandlerAccessor;
+import org.spongepowered.common.bridge.world.ServerWorldBridge;
 import org.spongepowered.common.bridge.world.WorldBridge;
 import org.spongepowered.common.data.provider.DataProviderRegistry;
 import org.spongepowered.common.event.tracking.PhaseContext;
@@ -85,6 +90,7 @@ import org.spongepowered.math.vector.Vector3i;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
+import java.time.Duration;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
@@ -300,5 +306,77 @@ public abstract class ServerWorldMixin_API extends WorldMixin_API<org.spongepowe
     public <E> Optional<E> get(int x, int y, int z, Key<? extends Value<E>> key) {
         final DataProvider<? extends Value<E>, E> dataProvider = DataProviderRegistry.get().getProvider(key, ServerLocation.class);
         return dataProvider.get(ServerLocation.of(this, new Vector3d(x, y, z)));
+    }
+
+    // WeatherUniverse
+
+    @Override
+    public Weather getWeather() {
+        if (this.shadow$isThundering()) {
+            return Weathers.THUNDER.get();
+        }
+        if (this.shadow$isRaining()) {
+            return Weathers.RAIN.get();
+        }
+        return Weathers.CLEAR.get();
+    }
+
+    @Override
+    public Duration getRemainingWeatherDuration() {
+        return Duration.of(this.api$getDurationInTicks(), TemporalUnits.MINECRAFT_TICKS);
+    }
+
+    private long api$getDurationInTicks() {
+        if (this.shadow$isThundering()) {
+            return this.worldInfo.getThunderTime();
+        }
+        if (this.shadow$isRaining()) {
+            return this.worldInfo.getRainTime();
+        }
+        if (this.worldInfo.getClearWeatherTime() > 0) {
+            return this.worldInfo.getClearWeatherTime();
+        }
+        return Math.min(this.worldInfo.getThunderTime(), this.worldInfo.getRainTime());
+    }
+
+    @Override
+    public Duration getRunningWeatherDuration() {
+        return Duration.of(this.worldInfo.getGameTime() - ((ServerWorldBridge) this).bridge$getWeatherStartTime(), TemporalUnits.MINECRAFT_TICKS);
+    }
+
+    @Override
+    public void setWeather(final Weather weather) {
+        Preconditions.checkNotNull(weather);
+        this.api$setWeather(weather, (300 + this.rand.nextInt(600)) * 20);
+    }
+
+    @Override
+    public void setWeather(final Weather weather, final Duration duration) {
+        Preconditions.checkNotNull(weather);
+        ((ServerWorldBridge) this).bridge$setPreviousWeather(this.getWeather());
+        final int ticks = (int) (duration.toMillis() / TemporalUnits.MINECRAFT_TICKS.getDuration().toMillis());
+        this.api$setWeather(weather, ticks);
+    }
+
+    public void api$setWeather(final Weather weather, final int ticks) {
+        if (weather == Weathers.CLEAR.get()) {
+            this.worldInfo.setClearWeatherTime(ticks);
+            this.worldInfo.setRainTime(0);
+            this.worldInfo.setThunderTime(0);
+            this.worldInfo.setRaining(false);
+            this.worldInfo.setThundering(false);
+        } else if (weather == Weathers.RAIN.get()) {
+            this.worldInfo.setClearWeatherTime(0);
+            this.worldInfo.setRainTime(ticks);
+            this.worldInfo.setThunderTime(ticks);
+            this.worldInfo.setRaining(true);
+            this.worldInfo.setThundering(false);
+        } else if (weather == Weathers.THUNDER.get()) {
+            this.worldInfo.setClearWeatherTime(0);
+            this.worldInfo.setRainTime(ticks);
+            this.worldInfo.setThunderTime(ticks);
+            this.worldInfo.setRaining(true);
+            this.worldInfo.setThundering(true);
+        }
     }
 }
